rebased and reworked PR
